move "main entry point" fns out from code.rs
We are working our way towards code.rs becoming
just the code for parsing expressions.
